text and float

+++ backend/db/migrations/20240801220558-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.FLOAT(9, 7),
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         lng: {
-          type: Sequelize.FLOAT(10, 7),
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         name: {
@@ -57,7 +57,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.FLOAT(7, 2),
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         createdAt: {
